+++ src/Cache/CacheableTrait.php
@@ -45,9 +45,6 @@ trait CacheableTrait
      *
      * @return bool
      *   Item is cached.
-     *
-     * @phpstan-ignore-next-line
-     * @throws \Psr\SimpleCache\InvalidArgumentException
      */
     protected function cacheSet(string $key, $value, $ttl = null): bool
     {
@@ -65,9 +62,6 @@ trait CacheableTrait
      *
      * @return bool
      *   Item is deleted.
-     *
-     * @phpstan-ignore-next-line
-     * @throws \Psr\SimpleCache\InvalidArgumentException
      */
     protected function cacheDelete(string $key): bool
     {
@@ -98,11 +92,8 @@ trait CacheableTrait
      *
      * @return mixed
      *   Cached value or default if no cache for the item.
-     *
-     * @phpstan-ignore-next-line
-     * @throws \Psr\SimpleCache\InvalidArgumentException
      */
-    protected function cacheGet(string $key, $default = null)
+    protected function cacheGet(string $key, mixed $default = null): mixed
     {
         return $this->hasCache()
             ? $this->cache->get($key, $default)