+++ tests/Client/Exception/InvalidResponseTest.php
@@ -5,15 +5,15 @@ declare(strict_types=1);
 namespace DigipolisGent\Tests\API\Client\Exception;
 
 use DigipolisGent\API\Client\Exception\InvalidResponse;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamInterface;
 
-/**
- * @covers \DigipolisGent\API\Client\Exception\InvalidResponse
- */
-class InvalidResponseTest extends TestCase
+#[CoversClass(InvalidResponse::class)]
+final class InvalidResponseTest extends TestCase
 {
     use ProphecyTrait;
 
@@ -26,8 +26,11 @@ class InvalidResponseTest extends TestCase
         $data = json_encode(['value' => uniqid('', true)], JSON_THROW_ON_ERROR);
         $statusCode = random_int(200, 500);
 
+        $stream = $this->prophesize(StreamInterface::class);
+        $stream->__toString()->willReturn($data);
+
         $response = $this->prophesize(ResponseInterface::class);
-        $response->getBody()->willReturn($data);
+        $response->getBody()->willReturn($stream->reveal());
         $response->getStatusCode()->willReturn($statusCode);
 
         $exception = InvalidResponse::fromResponse($response->reveal());