+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All Notable changes to `digipolisgent/api-client` package.
 
+## [Unreleased]
+
+### Updated
+
+* Update qa-php to 2.x.
+* Update psr/http-message to 1.x or 2.x.
+* Update psr/simple-cache to 1.x or 2.x or 3.x.
+
 ## [4.0.0]
 
 _Use version 4 of this module to switch over to OAuth authentication.